Time: 22 ms (8.51%), Space: 22.8 MB (73.84%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.cpp
@@ -27,7 +27,7 @@ public:
        
         while(low<=high){
             int mid=(low+high)/2;
-            int totalH=calculateTotalHours(piles,mid);
+            long long totalH=calculateTotalHours(piles,mid);
             if(totalH<=h){
                
                 high=mid-1;
